How do I solve the FATAL:  role "HackerNews" does not exist error?
The postgres database is like a warehouse. Actually, it's like a company that specializes in warehouses.  The company keeps accounts for specific people who want to rent warehouses. A 'role' is like an account for a specific person.  Rails assuming that a role called hackernews existing in postgres is like your manager assuming that you've opened a new account with the warehousing company in the name of your company.

I haven't opened a new account,and don't know how to.  So I just delete that assumption, and create an account in my own personal name and open three new warehouses called Hacker News for the stuff my app needs.