How do I add a navbar to the site?
Using the link_to helper.

The link_to helper needs to know the name of the link, the route it is linking to, and sometimes the http verb.  Specifically, you need the http verb when the request is not a gets.

Also, devise has a helper called current_user that you can use to see if anyone is logged in.

Navbar complete!